+++ tests/Migration/M190226185953TestMariadbMigration.php
@@ -15,7 +15,7 @@ class M190226185953TestMariadbMigration extends Mysql\Migration
      */
     public function safeUp()
     {
-        if (getenv('DB_TYPE') === 'mariadb') {
+        if ($this->getDb()->getDriverName() === 'mysql' && getenv('DB_TYPE') === 'mariadb') {
             $this->createTable('test_table', [
                 'enum_column_from_array' => $this->enum(['1', '2', '3', '4']),
                 'enum_column_from_enum' => $this->enum(TestEnum::class)
@@ -28,7 +28,7 @@ class M190226185953TestMariadbMigration extends Mysql\Migration
      */
     public function safeDown()
     {
-        if (getenv('DB_TYPE') === 'mariadb') {
+        if ($this->getDb()->getDriverName() === 'mysql' && getenv('DB_TYPE') === 'mariadb') {
             $this->dropTable('test_table');
         }
     }
